+++ src/main/java/eu/minemania/watson/chat/command/FakeCommandSource.java
@@ -7,6 +7,7 @@ import java.util.stream.Collectors;
 
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.server.command.CommandOutput;
 import net.minecraft.server.command.ServerCommandSource;
 
 public class FakeCommandSource extends ServerCommandSource
@@ -16,7 +17,7 @@ public class FakeCommandSource extends ServerCommandSource
 
     public FakeCommandSource(ClientPlayerEntity player)
     {
-        super(player, player.getPos(), player.getRotationClient(), null, 0, player.getName().getString(), player.getDisplayName(), null, player);
+        super(CommandOutput.DUMMY, player.getPos(), player.getRotationClient(), null, 0, player.getName().getString(), player.getDisplayName(), null, player);
     }
 
     @Override