+++ src/main/java/eu/minemania/watson/event/RenderHandler.java
@@ -5,7 +5,10 @@ import eu.minemania.watson.render.OverlayRenderer;
 import eu.minemania.watson.render.WatsonRenderer;
 import fi.dy.masa.malilib.interfaces.IRenderer;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.client.render.Camera;
+import net.minecraft.client.render.Fog;
+import net.minecraft.client.render.Frustum;
+import net.minecraft.util.profiler.Profiler;
 import org.joml.Matrix4f;
 
 public class RenderHandler implements IRenderer
@@ -18,14 +21,14 @@ public class RenderHandler implements IRenderer
     }
 
     @Override
-    public void onRenderWorldLast(Matrix4f matrix4f, Matrix4f projMatrix)
+    public void onRenderWorldLastAdvanced(Matrix4f posMatrix, Matrix4f projMatrix, Frustum frustum, Camera camera, Fog fog, Profiler profiler)
     {
         MinecraftClient mc = MinecraftClient.getInstance();
 
         if (Configs.Generic.ENABLED.getBooleanValue() && mc.world != null && mc.player != null)
         {
             OverlayRenderer.renderOverlays(mc);
-            WatsonRenderer.getInstance().piecewiseRenderEntities(mc);
+            WatsonRenderer.getInstance().piecewiseRenderEntities(mc, profiler, fog);
         }
     }
 }