+++ src/main/java/eu/minemania/watson/gui/Icons.java
@@ -3,6 +3,7 @@ package eu.minemania.watson.gui;
 import eu.minemania.watson.Reference;
 import fi.dy.masa.malilib.gui.interfaces.IGuiIcon;
 import fi.dy.masa.malilib.render.RenderUtils;
+import net.minecraft.client.gui.DrawContext;
 import net.minecraft.util.Identifier;
 
 public enum Icons implements IGuiIcon
@@ -54,7 +55,7 @@ public enum Icons implements IGuiIcon
     }
 
     @Override
-    public void renderAt(int x, int y, float zLevel, boolean enabled, boolean selected)
+    public void renderAt(int x, int y, float zLevel, boolean enabled, boolean selected, DrawContext drawContext)
     {
         RenderUtils.drawTexturedRect(x, y, this.u, this.v, this.w, this.h, zLevel);
     }