+++ src/main/java/eu/minemania/watson/render/WatsonRenderer.java
@@ -8,8 +8,9 @@ import eu.minemania.watson.db.BlockEditSet;
 import eu.minemania.watson.selection.EditSelection;
 import fi.dy.masa.malilib.render.RenderUtils;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.render.BackgroundRenderer;
+import net.minecraft.client.render.Fog;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.profiler.Profiler;
 import org.joml.Matrix4fStack;
 
 public class WatsonRenderer
@@ -23,7 +24,7 @@ public class WatsonRenderer
         return INSTANCE;
     }
 
-    public void piecewiseRenderEntities(MinecraftClient mc)
+    public void piecewiseRenderEntities(MinecraftClient mc, Profiler profiler, Fog fog)
     {
         if (this.mc == null)
         {
@@ -31,9 +32,8 @@ public class WatsonRenderer
         }
         if (Configs.Generic.DISPLAYED.getBooleanValue() && this.mc.getCameraEntity() != null && Configs.Outlines.OUTLINE_SHOWN.getBooleanValue())
         {
-            this.mc.getProfiler().push("watson_entities");
-            float fogStart = RenderSystem.getShaderFogStart();
-            BackgroundRenderer.clearFog();
+            profiler.push(() -> "watson_entities");
+            RenderSystem.setShaderFog(Fog.DUMMY);
             EditSelection selection = DataManager.getEditSelection();
             BlockEditSet edits = selection.getBlockEditSet();
             Matrix4fStack matrixStack = RenderSystem.getModelViewStack();
@@ -50,7 +50,6 @@ public class WatsonRenderer
             Vec3d cameraPos = this.mc.gameRenderer.getCamera().getPos();
 
             matrixStack.translate((float) -cameraPos.getX(), (float) -cameraPos.getY(), (float) -cameraPos.getZ());
-            RenderSystem.applyModelViewMatrix();
             edits.drawOutlines();
             edits.drawVectors();
             selection.drawSelection();
@@ -63,8 +62,8 @@ public class WatsonRenderer
             RenderSystem.enableCull();
 
             matrixStack.popMatrix();
-            RenderSystem.setShaderFogStart(fogStart);
-            this.mc.getProfiler().pop();
+            RenderSystem.setShaderFog(fog);
+            profiler.pop();
         }
     }
 }